requirements.txt
watchdog
pyyaml
nmcli